refactor: 중복 로직 함수화 #95
