+++ server/apps/api/src/community/helper/checkUserIsInCommunity.ts
@@ -1,0 +1,2 @@
+export const IsUserInCommunity = (user, communityId) =>
+  (user.communities ?? false) && Array.from(user.communities.keys()).includes(communityId);
