+++ server/apps/api/src/community/helper/index.ts
@@ -1,0 +1,2 @@
+export * from './checkUserIsInCommunity';
+export * from './makeCommunityObj';
