+++ server/apps/api/src/community/helper/makeCommunityObj.ts
@@ -1,0 +1,8 @@
+export const makeCommunityObj = (community_id: string) => {
+  const newCommunity = {};
+  newCommunity[`communities.${community_id}`] = {
+    _id: community_id,
+    channels: {},
+  };
+  return newCommunity;
+};
